Deleting items from the day's list
This commit introduces functionality to the delete button. The database
process is like this: the scrollview on the delete screen has buttons
with IDs that are the exact ROWIDs of the entry in the database. The
buttons have a template in the kv file that defines a lookup based on
it's ID as the rowid. The journal is regenerated after to clear old
entries.